requirements.txt
fastapi
uvicorn
pydantic
python-dotenv
google-generativeai>=0.7.2